requirements.txt
accelerate==1.10.0
aiofiles==24.1.0
aiohappyeyeballs @ file:///home/conda/feedstock_root/build_artifacts/aiohappyeyeballs_1741775197943/work
aiohttp @ file:///home/conda/feedstock_root/build_artifacts/aiohttp_1753804962438/work
aiosignal @ file:///home/conda/feedstock_root/build_artifacts/aiosignal_1751626463503/work
annotated-types==0.7.0
anyio==4.10.0
async-timeout @ file:///home/conda/feedstock_root/build_artifacts/async-timeout_1733235340728/work
attrs @ file:///home/conda/feedstock_root/build_artifacts/attrs_1741918516150/work
av @ file:///home/conda/feedstock_root/build_artifacts/av_1751744976050/work
bitsandbytes==0.47.0
Brotli @ file:///home/conda/feedstock_root/build_artifacts/brotli-split_1749229842835/work
certifi @ file:///home/conda/feedstock_root/build_artifacts/certifi_1754231422783/work/certifi
cffi @ file:///home/conda/feedstock_root/build_artifacts/cffi_1725560520483/work
charset-normalizer @ file:///home/conda/feedstock_root/build_artifacts/charset-normalizer_1754767332901/work
click==8.2.1
colorama @ file:///home/conda/feedstock_root/build_artifacts/colorama_1733218098505/work
decord==0.6.0
diffusers==0.34.0
dill @ file:///home/conda/feedstock_root/build_artifacts/dill_1690101045195/work
exceptiongroup==1.3.0
fastapi==0.116.1
ffmpy==0.6.1
filelock @ file:///home/conda/feedstock_root/build_artifacts/filelock_1755216263872/work
frozenlist @ file:///home/conda/feedstock_root/build_artifacts/frozenlist_1752167142843/work
fsspec==2024.3.1
future @ file:///home/conda/feedstock_root/build_artifacts/future_1738926421307/work
gitdb==4.0.12
GitPython==3.1.45
gradio==5.42.0
gradio_client==1.11.1
groovy==0.1.2
h11==0.16.0
h2 @ file:///home/conda/feedstock_root/build_artifacts/h2_1738578511449/work
hf-xet @ file:///home/conda/feedstock_root/build_artifacts/bld/rattler-build_hf-xet_1754484932/work
hpack @ file:///home/conda/feedstock_root/build_artifacts/hpack_1737618293087/work
httpcore==1.0.9
httpx==0.28.1
huggingface_hub @ file:///home/conda/feedstock_root/build_artifacts/huggingface_hub_1754700484541/work
hyperframe @ file:///home/conda/feedstock_root/build_artifacts/hyperframe_1737618333194/work
idna @ file:///home/conda/feedstock_root/build_artifacts/idna_1733211830134/work
imageio==2.37.0
imageio-ffmpeg==0.6.0
importlib_metadata @ file:///home/conda/feedstock_root/build_artifacts/bld/rattler-build_importlib-metadata_1747934053/work
Jinja2==3.1.6
markdown-it-py==4.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
mkl-service==2.4.0
mpmath==1.3.0
multidict @ file:///home/conda/feedstock_root/build_artifacts/multidict_1751310558090/work
multiprocess @ file:///home/conda/feedstock_root/build_artifacts/multiprocess_1695458832722/work
networkx==3.4.2
ninja==1.13.0
numpy @ file:///home/conda/feedstock_root/build_artifacts/numpy_1707225380409/work/dist/numpy-1.26.4-cp310-cp310-linux_x86_64.whl#sha256=51131fd8fc130cd168aecaf1bc0ea85f92e8ffebf211772ceb16ac2e7f10d7ca
nvidia-cublas-cu12==12.8.4.1
nvidia-cuda-cupti-cu12==12.8.90
nvidia-cuda-nvrtc-cu12==12.8.93
nvidia-cuda-runtime-cu12==12.8.90
nvidia-cudnn-cu12==9.10.2.21
nvidia-cufft-cu12==11.3.3.83
nvidia-cufile-cu12==1.13.1.3
nvidia-curand-cu12==10.3.9.90
nvidia-cusolver-cu12==11.7.3.90
nvidia-cusparse-cu12==12.5.8.93
nvidia-cusparselt-cu12==0.7.1
nvidia-nccl-cu12==2.27.3
nvidia-nvjitlink-cu12==12.8.93
nvidia-nvtx-cu12==12.8.90
optimum @ file:///home/conda/feedstock_root/build_artifacts/optimum_1753924024694/work
optimum-quanto==0.2.7
orjson==3.11.2
packaging @ file:///home/conda/feedstock_root/build_artifacts/bld/rattler-build_packaging_1745345660/work
pandas @ file:///home/conda/feedstock_root/build_artifacts/pandas_1752081716429/work
peft==0.17.0
pillow @ file:///home/conda/feedstock_root/build_artifacts/pillow_1751481981736/work
pillow_heif==1.1.0
platformdirs==4.3.8
propcache @ file:///home/conda/feedstock_root/build_artifacts/propcache_1744524934684/work
protobuf @ file:///home/conda/feedstock_root/build_artifacts/protobuf_1751668305329/work/bazel-bin/python/dist/protobuf-6.31.1-cp310-abi3-linux_x86_64.whl#sha256=defdfb61601fe45e5f630c511e9db761b2d454cda73e7e10af9b4112431a5435
psutil==7.0.0
pyarrow==14.0.2
pyarrow-hotfix==0.7
pycparser @ file:///home/conda/feedstock_root/build_artifacts/bld/rattler-build_pycparser_1733195786/work
pydantic==2.11.7
pydantic_core==2.33.2
pydub==0.25.1
Pygments==2.19.2
PySocks @ file:///home/conda/feedstock_root/build_artifacts/pysocks_1733217236728/work
python-dateutil @ file:///home/conda/feedstock_root/build_artifacts/bld/rattler-build_python-dateutil_1751104122/work
python-multipart==0.0.20
pytz @ file:///home/conda/feedstock_root/build_artifacts/pytz_1742920838005/work
PyYAML @ file:///home/conda/feedstock_root/build_artifacts/pyyaml_1737454647378/work
quanto==0.2.0
regex @ file:///home/conda/feedstock_root/build_artifacts/regex_1753984087656/work
requests @ file:///home/conda/feedstock_root/build_artifacts/requests_1749498106507/work
rich==14.1.0
ruff==0.12.9
safehttpx==0.1.6
safetensors @ file:///home/conda/feedstock_root/build_artifacts/safetensors_1754674849140/work
semantic-version==2.10.0
sentencepiece @ file:///home/conda/feedstock_root/build_artifacts/sentencepiece-split_1751790858326/work/python
sentry-sdk==2.35.0
shellingham==1.5.4
six @ file:///home/conda/feedstock_root/build_artifacts/bld/rattler-build_six_1753199211/work
smmap==5.0.2
sniffio==1.3.1
starlette==0.47.2
sympy==1.14.0
tokenizers @ file:///home/conda/feedstock_root/build_artifacts/tokenizers_1753767998928/work/bindings/python
tomlkit==0.13.3
torch==2.8.0
torchvision==0.23.0
tqdm @ file:///home/conda/feedstock_root/build_artifacts/tqdm_1735661334605/work
transformers @ file:///home/conda/feedstock_root/build_artifacts/transformers_1755135050945/work
triton==3.4.0
typer==0.16.0
typing-inspection==0.4.1
typing_extensions @ file:///home/conda/feedstock_root/build_artifacts/bld/rattler-build_typing_extensions_1751643513/work
tzdata @ file:///home/conda/feedstock_root/build_artifacts/python-tzdata_1742745135198/work
urllib3 @ file:///home/conda/feedstock_root/build_artifacts/urllib3_1750271362675/work
uvicorn==0.35.0
wandb==0.21.1
websockets==15.0.1
xxhash @ file:///home/conda/feedstock_root/build_artifacts/python-xxhash_1740594790928/work
yarl @ file:///home/conda/feedstock_root/build_artifacts/yarl_1749554822108/work
zipp @ file:///home/conda/feedstock_root/build_artifacts/zipp_1749421620841/work
zstandard==0.23.0
